fix(kalshi): reconcile format/narrowing + cap-truncation test
- cast qsizetype out.size() to int for the %d qWarning format spec
  (was -Wformat under -Werror in CI)
- cast id.indexOf('-') to int at the assignment site to silence the
  qsizetype->int narrowing warning
- drop unused QLoggingCategory include in favor of QtGlobal
- add desired_truncates_to_cap test exercising the cap/truncation path
- rename reconcile_never_touches_foreign_held to
  reconcile_empty_delta_when_desired_equals_held (it was testing the
  identity case, not foreign-held scoping)

+++ openmarketterminal-qt/src/services/prediction/kalshi/Kalshi15mReconcile.cpp
@@ -1,5 +1,5 @@
 #include "services/prediction/kalshi/Kalshi15mReconcile.h"
-#include <QLoggingCategory>
+#include <QtGlobal>
 
 namespace kalshi15m {
 
@@ -9,7 +9,7 @@ QStringList desired_subscriptions(const QVector<pred::PredictionMarket>& markets
     QSet<QString> seen;
     for (const auto& m : markets) {
         const QString id = m.key.market_id;
-        const int dash = id.indexOf('-');
+        const int dash = static_cast<int>(id.indexOf('-'));
         const QString family = dash < 0 ? id : id.left(dash);
         if (!families.contains(family) || seen.contains(id) || id.isEmpty())
             continue;
@@ -18,7 +18,7 @@ QStringList desired_subscriptions(const QVector<pred::PredictionMarket>& markets
     }
     if (cap > 0 && out.size() > cap) {
         qWarning("kalshi15m: %d open markets exceeds cap %d; capping",
-                 out.size(), cap);
+                 static_cast<int>(out.size()), cap);
         out = out.mid(0, cap);
     }
     return out;
